create abstract
